do not look for pack to install in already installed pack
Previously, the install-gamepacks.sh script was looking for all
subdirectories in games/ directory for *.games directory to
copy them inside games/ directory.

So, on first run,
it was looking for games/XonoticPack/xonotic.game
to copy it as games/xonotic.game

Then on second run,
it was looking for games/XonoticPack/xonotic.game
and also for games/xonotic.game/xonotic.game
which is silly.

Now the install-gamepacks.sh script is looking for games/*Pack
subdirectories only, it means all gamepacks must end with a
capitalized Pack suffix, which is already the case and is defined
in download-gamepacks.sh after all.