Fix deprecated pango code always being used in glfont.cpp
The version check to see if the non-deprecated API is available,
PANGO_VERSION_CHECK(1,22,0), always returns false if pango-features.h
isn't included. This is because PANGO_VERSION_MAJOR, _MINOR, _MICRO
are not defined and are evaluated as 0 in the preprocessor code.

This commit fixes compiling on Debian 8 using libpango1.0-dev 1.36.8.
It was failing because deprecated function pango_ft2_get_context()
was not available (hidden in header file by PANGO_DISABLE_DEPRECATED).