restore “open userGamePath/maps/” behavior when LastMapFolder unset
The former behavior when LastMapFolder was not implemented
was to try to open maps/ dir in userGamePath.

When LastMapFolder code was written the code only looked
for userGamePath, this is restoring that behavior,
mimicking the code for model opening that does the
same with userGamePath/models.

As an extra, it also fallback on userGamePath is maps/
is missing.