floating window: use HINT_UTILITY on Windows but normal window everywhere else, ref !140 !147
GDK_WINDOW_TYPE_HINT_UTILITY would have been a good option for any OS but KDE is known to be broken (such window is not put in focus).

We must use GDK_WINDOW_TYPE_HINT_UTILITY on Windows to be sure the window is resizeable and to avoid the infinite minimize-restore loop.

<3 @SpiKe